requirements.txt
streamlit
pandas
numpy
openai
requests
python-dotenv
PyPDF2
docx2txt
python-docx
duckduckgo-search
beautifulsoup4
urllib3
certifi
pandera
pytest
scikit-learn
scipy
spacy
textstat